+++ src/shift_left/tests/ut/core/debug_ut.py
@@ -4,10 +4,10 @@ import os
 import pathlib
 import json
 import time
-
-#os.environ["CONFIG_FILE"] = str(pathlib.Path(__file__).parent.parent.parent / "config-ccloud.yaml")
-#os.environ["PIPELINES"] = str(pathlib.Path(__file__).parent.parent.parent / "data/flink-project/pipelines")
-os.environ["CONFIG_FILE"]= os.getenv("HOME") + ".shift_left/config-dev.yaml"
+import yaml
+os.environ["CONFIG_FILE"] = str(pathlib.Path(__file__).parent.parent.parent / "config-ccloud.yaml")
+os.environ["PIPELINES"] = str(pathlib.Path(__file__).parent.parent.parent / "data/flink-project/pipelines")
+#os.environ["CONFIG_FILE"]= os.getenv("HOME") + ".shift_left/config-dev.yaml"
 
 from shift_left.core.utils.app_config import get_config, shift_left_dir, logger
 from shift_left.core.models.flink_statement_model import ( 
@@ -21,7 +21,7 @@ from shift_left.core.pipeline_mgr import PIPELINE_JSON_FILE_NAME
 import shift_left.core.deployment_mgr as dm
 import shift_left.core.test_mgr as test_mgr
 import shift_left.core.table_mgr as table_mgr
-from shift_left.core.utils.file_search import build_inventory
+from shift_left.core.utils.file_search import get_or_build_inventory, get_table_ref_from_inventory, FlinkTableReference, from_pipeline_to_absolute
 import shift_left.core.deployment_mgr as deployment_mgr
 from ut.core.BaseUT import BaseUT
 
@@ -83,44 +83,28 @@ class TestDebugUnitTests(BaseUT):
         )
         return ComputePoolList(pools=[pool_1])
     
-    @patch('shift_left.core.deployment_mgr.statement_mgr.post_flink_statement')
-    @patch('shift_left.core.deployment_mgr.statement_mgr.drop_table')
-    @patch('shift_left.core.deployment_mgr.statement_mgr.delete_statement_if_exists')
-    def test_deploy_product_using_parallel(self, 
-                                           mock_delete, 
-                                           mock_drop,
-                                           mock_post):
-        def _drop_table(table_name: str, compute_pool_id: str) -> str:
-            print(f"@@@@ drop_table {table_name} {compute_pool_id}")
-            logger.info(f"@@@@ drop_table {table_name} {compute_pool_id}")
-            time.sleep(1)
-            return "deleted"
-        
-        def _post_flink_statement(compute_pool_id: str, statement_name: str, sql_content: str) -> Statement:
-            print(f"\n@@@@ post_flink_statement {compute_pool_id} {statement_name} {sql_content}")
-            logger.info(f"@@@@ post_flink_statement {compute_pool_id} {statement_name}\n {sql_content}")
-            time.sleep(1)
-            if "ddl" in statement_name:
-                return self._create_mock_statement(name=statement_name, status_phase="COMPLETED")
-            return self._create_mock_statement(name=statement_name, status_phase="RUNNING")
-        
-        def _delete_statement(statement_name: str):
-            print(f"@@@@ delete statement {statement_name}")
-            logger.info(f"@@@@ delete statement {statement_name}")
-            time.sleep(1)
-            return "deleted"
-        config=get_config()
-        mock_delete.side_effect = _delete_statement
-        mock_drop.side_effect = _drop_table
-        mock_post.side_effect = _post_flink_statement
+    
+    def test_create_yaml_file(self):
+        inventory_path = os.path.join(os.getenv("PIPELINES"),)
+        table_inventory = get_or_build_inventory(inventory_path, inventory_path, False)
+        table_ref: FlinkTableReference = get_table_ref_from_inventory(table_name="fct_user_per_group", inventory=table_inventory)
+        test_folder_path = from_pipeline_to_absolute(table_ref.table_folder_name) + "/tests"
+        test_definition = test_mgr._add_test_files(table_to_test_ref=table_ref, 
+                                                tests_folder=test_folder_path,
+                                                table_inventory=table_inventory,
+                                                create_csv=False,
+                                                nb_test_cases=1,
+                                                use_ai=False)
 
-        deployment_mgr.build_deploy_pipelines_from_product(product_name="qx", 
-                                                           inventory_path=self.inventory_path, 
-                                                           execute_plan=True,
-                                                           force_ancestors=True,
-                                                           sequential=False)
-
-   
+        assert test_definition is not None
+        
+        yaml_file = os.path.join(test_folder_path, '..', "fct_user_per_group_1.yaml")
+        with open(yaml_file, "r") as f:
+            yaml_content = yaml.load(f, Loader=yaml.FullLoader)
+        assert yaml_content is not None
+        assert "fct_user_per_group" in yaml_content
+ 
+        
 
 if __name__ == '__main__':
     unittest.main()